move VehiclePosition to separate file
