+++ libosmscout-client-qt/include/osmscout/NavigationModel.h
@@ -30,80 +30,6 @@
 namespace osmscout {
 
 /**
- * \ingroup QtAPI
- *
- * Object aggregating estimated data about vehicle during navigation.
- */
-class OSMSCOUT_CLIENT_QT_API VehiclePosition: public QObject
-{
-  Q_OBJECT
-
-  Q_PROPERTY(double   lat       READ getLat)
-  Q_PROPERTY(double   lon       READ getLon)
-  Q_PROPERTY(double   bearing   READ getBearingRadians)
-
-public:
-  inline VehiclePosition(QObject *parent = 0) : QObject(parent)
-  {}
-
-  inline VehiclePosition(const Vehicle &vehicle,
-                         const PositionAgent::PositionState &state,
-                         const GeoCoord &coord,
-                         const std::shared_ptr<Bearing> &bearing,
-                         const std::shared_ptr<GeoCoord> &nextStepCoord,
-                         QObject *parent = 0):
-      QObject(parent), vehicle(vehicle), state(state), coord(coord), bearing(bearing), nextStepCoord(nextStepCoord)
-  {}
-
-  inline VehiclePosition & operator=(const VehiclePosition &o)
-  {
-    vehicle=o.vehicle;
-    state=o.state;
-    coord=o.coord;
-    bearing=o.bearing;
-    nextStepCoord=o.nextStepCoord;
-    return *this;
-  }
-
-  inline double getLat() const
-  {
-    return coord.GetLat();
-  }
-
-  inline double getLon() const
-  {
-    return coord.GetLon();
-  }
-
-  inline std::shared_ptr<Bearing> getBearing() const
-  {
-    return bearing;
-  }
-
-  inline double getBearingRadians() const
-  {
-    return bearing ? bearing->AsRadians() : 0;
-  }
-
-  inline std::shared_ptr<GeoCoord> getNextStepCoord() const
-  {
-    return nextStepCoord;
-  }
-
-  inline PositionAgent::PositionState getState() const
-  {
-    return state;
-  }
-
-private:
-  Vehicle vehicle;
-  PositionAgent::PositionState state;
-  GeoCoord coord;
-  std::shared_ptr<Bearing> bearing;
-  std::shared_ptr<GeoCoord> nextStepCoord;
-};
-
-/**
  * Model providing navigation functionality to QML.
  * Main logic sits in osmscout::Navigation class and its Qt wrapper NavigationModule.
  *
